@@ -6,12 +6,11 @@ using com.gameon.domain.Models.MappingProfiles;
 using GameOnApi.CustomExceptionMiddleware;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Swashbuckle.AspNetCore.Swagger;
+using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
 
-[assembly: ApiController]
 namespace GameOnApi
 {
     public class Startup
@@ -28,6 +27,7 @@ namespace GameOnApi
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
+            services.AddControllers().AddNewtonsoftJson();
             services.AddCors(options =>
             {
                 options.AddPolicy(MyAllowSpecificOrigins,
@@ -53,11 +53,9 @@ namespace GameOnApi
             IMapper mapper = mappingConfig.CreateMapper();
             services.AddSingleton(mapper);
 
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
-
             services.AddSwaggerGen(c =>
             {
-                c.SwaggerDoc("v1", new Info { Title = "Game On API", Version = "v1" });
+                c.SwaggerDoc("v1", new OpenApiInfo { Title = "Game On API", Version = "v1" });
             });
 
             // Add services in data layer via ServiceManager and the domain layer
@@ -71,7 +69,7 @@ namespace GameOnApi
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -82,13 +80,12 @@ namespace GameOnApi
                 // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
                 app.UseHsts();
             }
-
-            app.UseCors(MyAllowSpecificOrigins);
+            
             app.UseSwagger();
             // Enable middleware to serve swagger-ui (HTML, JS, CSS, etc.), specifying the Swagger JSON endpoint.
             app.UseSwaggerUI(c =>
             {
-                c.SwaggerEndpoint("/swagger/v1/swagger.json", "Game On API");
+                c.SwaggerEndpoint("/swagger/v1/swagger.json", "Game On API v1");
                 // Set RoutePrefix to an empty string to serve swagger UI as the app's root
                 c.RoutePrefix = string.Empty;
             });
@@ -99,7 +96,16 @@ namespace GameOnApi
             app.UseStaticFiles();
             app.UseCookiePolicy();
 
-            app.UseMvc();
+            app.UseRouting();
+            app.UseCors(MyAllowSpecificOrigins);
+
+            //app.UseAuthentication();
+            //app.UseAuthorization();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllers();
+            });
         }
     }
 }
